feat: implement color and the framebuffer
this commit adds a simple color definition with some colorformats
supported by sdl, as well as some conversion utilities and a 256 color
palette.

it also implements a framebuffer that deals with all the math to map
x, y coordinates and color values to the underlying bytearray so that
we don't have to deal with those calculations later (it could be
refactored to avoid the duplication in the two methods... next time).

updated the sdl test to use the newly defined constructs and it
behaves the same as before.

next up I'll make the sdl code more reusable and will start
implementing basic 2D primitives.