fix: correct window to framebuffer scaling
when creating a sketch the intended size of the sketch (that is, its
width and height) is intended to be the size of the drawable
surface (the framebuffer size) and not exactly the size of the window
itself.

== the issue

the way as implemented before, we'd pass a size into the window along
with a scaling factor and the window would scale the framebuffer.

thiw would mean that when using the window size as coordinates for
anything - as we do in the canvas bounds example - and the scaling
factor is > 1, the coordinates end up incorrectly mapped (specially
once we move beyond the framebuffer defined bounds).

For example:

when creating sketch with size 640 x 480 (w x h) and scaling factor of
2, the framebuffer size is 320 x 240.

at the sketch level, the width of the sketch is that of the window so
it would be reasonable to address a pixel location at 635, 475 (x, y)
since the sketch size is 640, 480... except these coordinates would
fall outside the bounds of the framebuffer.

one could force the window client (the sketch) to account for this,
but then it makes things very strange when scaling up in sdl.

== the fix (???)

we make the sketch size match to the framebuffer and then 'scale' the
window. this any point within the sketch falls on the framebuffer.

the only thing to know is that the scale will affect the actual window
size.

tbh, though, I have a feeling input event coordinates could have some
dragon waiting for me later... need to think about this a bit more,
but so far it seems to work ok.