test: add and tweak lines example to show bug in line implementation
constraining the first line example to leave a bit of margin surfaced
a bug in the existing line implementation. it seems to be related to
the slope of the line.

I went ahead and wrote another - less random - example that uses lines
and boom, it's more obvious.

Next up, figure out what's going on.