refactor: cleanup examples and simplify how to run them
removed lots of old copy pasta as I wrote some of the examples and put
them all under one `examples` script to simplify things.

could've gone with some nicer curses menu and stuff, but want to keep
dependencies to a minimum (so far just SDL, which is big enough as it
is).