feat: implement Point protocol from the blue book
This commit extends the definition of a Point (and starts off with
Rectangle as well) to be used at a later point (lol) in when playing
with some more complex graphics operations. It follows the definition
outline in "Smalltalk-80: The Language and its Implementation" by
Adele Goldberg and David Robson (the blue book), but somewhat crammed
into python style (a lot of it would translate so nicely to ruby).

It's nice that the chapter that describes these protocols has some
examples that can be used as test cases (they are - however - not
exhaustive!), but in some cases python's semantics necessitate a
few changes (rounding, floating point comparisons).

That's all for today. Will be a couple of busy days the rest of the
week so not sure if I'll be able to swing any more changes before
the weekend, but I'd like to move onto the functionality that uses
these geometric constructs.