feat: implement the task's execution loop (sans eval).
- tasks are executed by passing a word to their start method.
- tasks then get the word's definition and load it onto the callstack.
- tasks then run by poping definitions from the callstack into the
  callframe and executing them.
- execution of the callframe involves walking through the values of
  the current quotation in the callframe and evaluating them.
  - if the values are words they are executed.
  - else values are pushed onto the datastack so that words can
    operate on them.

The last points in the above remain to be implemented and I'm not sure
that Task is the right name for this construct. Perhaps interpreter or
procedure or process... I got the Task name from the Oberon book as it
seemed to have a nice way of cataloging processes... just not sure
about the name. We'll see.