feat: define how a task is bootstrapped/loaded and executed.
a lot of this process comes from an earlier attempt at this language
that I worked on during the holidays and it draws heavily from some
of the earlier Java implementations of the Factor programming
language (somewhere around 0.66 or 0.67 maybe), where this is the
process used to bootstrap the interpreter.

the `call` word takes a quotation on top of the stack and sets it
as the current callframe, if there's already a callframe being
evaluated it is pushed onto the callstack to resume later.

the run (maybe it should be execute) method advances through the
values in the callframe quotation and evaluates them. if it reaches
the end of the callframe it gets the next one from the callstack and
if there's no more quotations in the callstack, it terminates
execution.

the idea here is that a task is started by supplying a word to the
task's `start` method, the task then pushes the word's definition onto
the datastack and executes `call` which will set it as the current
callframe to start the process above.