refactor: rename task to interpreter
the task was basically the interpreter from `ruin` (my previous
attempt at this) it seemed weird to name it a task. There's still a
concept of a Task or Proc with similar taxonomy as in Oberon, but
these will hold an interpreter that is doing the work and allows the
Task/Proc to also have other components unrelated to the main
interpreter execution.

I also considered:
- ExecutionContext
- VMContext
- InterpreterContext
- Runtime

naming things is hard!