feat: add namespaces and support parse words
parse words are executed by the parser at parse time and not added to
the parse tree. the idea is that when the parser sees a parse word it
sets itself on the interpreter's global namespace and then invokes the
parse word.

this required adding namespace support, though right now only a global
namespace is being used.

used this to implement `IN:` which is used to set the current vocabulary.