feat: add user defined words
sigh... this took much longer than I expected it to take, though to be
honest I'm not really sure what was I expecting, this was the same
story with `rtkm` and I never got here with `ruin`. and yet, I
expected it to go smoother.

a word can be defined as `: square dup * ;` and is parsed thus:

- the `:` word is a parse word so the parser immediately executes
  (as opposed as parsing it and adding it to the callstack for
  evaluation time).

- the next value - the name of the word - is read in raw mode and
  pushed onto the datastack.

- the definer word (`define-word`) is fetched from the loaded
  vocabularies and pushed onto the stack as well.

- an empty quotation is pushed onto the stack to act as an accumulator
  for the parser. the parser will keep on parsing values off the
  source and appending them to this accumulator, these values make up
  the word's definition.

- the `;` (`endefinition`) is also a parse word, so the parser
  executes it.

- the top two value of the stack are swapped so that the `define-word`
  can be removed.

- `define-word` is removed and evaluated.

- `define-word` gets the word definition (a quotation) and its name
   and creates the new word.

- the parser is not using the earlier accumulator quotation.

a note about `;` because its sole purpose is to evaluate the definer,
it can be used as a terminator delimiter for other parse words. The
parse word just needs to make sure to set up the stack so that `;` can
get at the definer word in the same way.

ran into quite a few issues that stem from how I've been building this
thing so I got no one to blame but myself. still, it feels like it's
coming along.

TODO:

code & pkg structure:
- move vocabs from primitives into vocab
- move stack out of primitives
- rename primitives to primitive
- move codebase from vm to interpreter
  - there's no reason for the machine to have access to it.
  - the indirection is a pain
  - with the addition of tasks I'll have to add locks

features:
- collections (assocs, tuples, quotations).
- module imports (USE: USING:)
- vocab order resolution
- ffi (specifically python ffi)
  - I've written this twice (`rtkm` and `ruin`) so I feel I have a
    much better feeling for what worked and didn't then.
- standard library:
  - control flow combinators
  - classes and multimethods (generic words and methods)
  - math
  - strings
  - io
  - network
  - http
- revisit the graphics work I had started in `yak.draw`.
- more tests.