refactor: extract vocabularies out of primitives
while the existing vocabularies are made out of primitives, it felt
confusing having them along with the primitive python code.

it still feels somewhat confusing, but the line I've drawn is this, if
they are python functions operating on the interpreter directly then
they go in `vocab` and if they are python functions oeprating on other
values they go in `primitives`.