refactor: remove vm altogeher (hal/os now)... clean up some things
tbh I'm not sure I'm entirely happy with the way things are shaping
up. I think it all stems from attempting to start with an idea for a
larger systems and then trying to fill it in a piece at a time instead
of going from a small kernel and grow it up...

so I'm back at the smaller kernel... a single execution process
interpreter, but with some clunky things like this codebase construct
that was meant to be something else and now it's just a collection of
vocabularies that could've been handled differently?

I'll have to go back and revisit that a bit as I expand on namespaces,
but for now I'm going to focus a bit more on the language syntax w/ a
single namespace.