feat: implement DEFINE: parse word.
with `DEFINE:` I'm now able to have the full vocabulary declared in a
`.yak` file (and - eventually - in a code database), even for
primitive words that are implemented in the host language (in this
case Python).

I had to compromise a little on the syntax to make it possible to also
add stack effects and comments/docstrings to these primitive
declarations in `yak` source code.

Also implemented a dummy parser for stack effects. Right now it just
eats through them and discards them, but it allows me to add
stack-effects in `yak` source now and then get the benefits when
proper stack effect parsing and checking is implemented.