feat: bootstrap vocabularies
with primitive and some updates to `vocab.bootstrap` I'm able now to
load more and more vocabs from `yak` source.

as I removed kernel from explicitly being loaded in the interpreter
some things broke because printing words are not defined in
`kernel.yak`, but the loading works.

had to fix a few bugs but it's all looking on the up and up.

tomorrow need to boostrap remaining vocabs, add printing, and look
into looping.