feat: add DEFER: parse word to allow forward declarations.
implemented `DEFER:` to allow "forward declarations" and recursive
words. it's a bit hacky and maybe not how it should be done, but so
far it works for the way `yak` is evaluated.

`DEFER:` reads the word and pushes it onto the stack along with an
empty quotation and then calls `define-compound` directly. this
effectively creates a noop word.

as long as the word is not called before parsing is complete (say
at parse time) it will work because eventually the word being
deferred would - ideally - be defined.

If that is not the case an error is thrown. There's no error handling,
but the plan is to then allow the user to define the word at runtime
and continue or retry evaluation (restarts!).

the main driver for this naive implementation of `DEFER:` is that I
needed it to implement iteration words! This may not seem as much, but
after banging my head against the wall over and over again trying to
use this implementation in `rtkm` and failing I'm happy it's working
well (for now).

the main issue in `rtkm` was that I just evaluated everything by
traversing the parsed quotation and evaluating values as I went, there
was no callstack and this made this `while` implementation fail in
multiple ways... basically recursion was borked... I ended up
implementing iteration words as python primitives.

then I dug through earlier implementations of factor and the docs for
the makovm and between them I got an idea of how to do this better
using a callstack (from factor) and a retainstack (from mako docs and
other forth texts). some of this went into `ruin` but that ended up
having other issues.

so I'm happy it's working now ... oh how do I know? I've tested it
with my first `yak` program.