feat: implement a simple repl (in yak).
changed `read` to take a prompt string as input, if I don't want a
prompt I can simply pass "" and be done with it. otherwise one needs
to do `">> " print read ;` which looks like it'd work except that
`print` adds a new line at it ends up looking like:

```
>>

```

and requiring a prompt for `read` is simpler right now that
implementing more printing words (I'll still need to do this, just not
right now).

also removed the previous noise in `io.yak`

here's the first `yak` repl session.
```
(yak) % yak
>> 3
>> 4
>> print
4
>> : put4 ( -- 4 ) 4 ;
>> put4
>> put4
>> put4
>> print
4
>> print
4
>> print
4
>> quit-repl
bye!
There was an error executing `throw`, error=There is no vocabulary with name=errors
```

need to add some words to print the stack for the *print* part of repl

there's also something about the implementation of `while` in that
when it the loop terminates it drops the predicate and body from the
datastack but the callstack still has:

```
[ >r [ call ] keep r> swap while ]
```

not good... time for some other approach lol.