fix: change `while` execute and repeat step if condition holds
before sleeping last night I grabbed pen and paper and worked through
my previous implementation again and noticed it was dropping the
predicate and the body of the loop if the condition didn't hold, but
then it always tried to execute the loop body and recurse. that last
step had to be conditional as well.

fix was fairly straightforward... use `if-else`.

```
(yak) % yak
>> 1
>> 2
>> 3
>> 4
>> : 4print print print print print ;
>> 4print
4
3
2
1
>> 5
>> quit-repl
bye!
(yak) %
```