feat: porting parser and scanner from interpreter
this is a bit where I start going off the path in "Crafting
Interpreters" (for now at least) as `yak` is not much like `(c)lox`
and I have to connecting my old parser and scanner from the
interpreter into a compiler to emit bytecode for the vm.

this will impact how the parser accumulated values and certainly how
to handle words and parse words.

getting to the point where I can do airthmetic with `yak` source will
be nice.