wip: start interpreter rewrite, lots of broken imports atm
start a new parser that keeps its own definition stack and exposes
the same protocol that syntax words expect of the interpreter. this is
so that syntax words can work at parse and runtime, at parse time they
operate on the parser state and at runtime they operate on the
interpreter state.

added a callframe class that holds onto the quotation with the code
and uses an instruction pointer to traverse this quotation. this will
help later with looping constructs and condition handling.

at the moment, this new interpreter module is pretty broken as
there are a lot of broken/missing imports, but I'm just trying to lay
down some ideas and lifting some of the things in the previous
interpreter that I felt worked ok.