exp: playing around with an http canvas?
some loose ideas in my head...

```
% curl -XPOST localhost:45133/exec -H "Content-Type: application/json" -d '{"cmd": "random-fill"}'
% curl -XPOST localhost:45133/exec -H "Content-Type: application/json" -d '{"cmd": "signal-exit"}'
```

some issues tho:

I think I messed up the scale or the fill... sure it's somethign
simple, but I'm tired.

the machine-server circular relationship is ... I don't like it and it
may be causing some issues when exiting.

need a good way to exit the server...