feat: implement Bresenham's line drawing using bitblt
implemented bresenham's using the bluebook's implementation as a
reference as well as some old squeak code. ran into some strange bugs
during implementation and had to refer to the old implementation of
bresenham's in fb.py to compare notes.

mostly works, but I think I've found an off by one error in the code
that copies bytes in and out of a form's bitmap (bytearray in this
case). 1 pixel squares do not render... I'll look at that later.

can't say for sure if I prefer this over the fb.py implementation in
terms of ergonomics, but I gotta say that this solves the line
thickness problem that originally sent me this way. Also, bitblt is
pretty versatile for all sorts of other operations (as soon as I
implement the rest of it) so hopefully I can get much more out of it
as I build on it.

next up, spend some time on an actual interpreter... however simple.