exp: implement a persistent object pseudo dsl
I like the idea of persistent objects/runtimes and orthogonal persistence in general. I understand all this comes with tradeoffs, but for a certain size or scope of projects it seems like a good fit... particularly thinking of very small personal programs/tools.

a while back I came accross charlesetc's [redpear and walnut](https://inclouds.space/blog/redpear/) and since I've been working in python a lot recently I started thinking what it'd take to implement something like [walnut](https://github.com/charlesetc/walnut/tree/main) in python.

a recent conversation (more like a rant) got me thinking about it, so I took a stab at a first pass of this for a small chunk this morning. it's not even half done, but seems promising in a way.

a few scattered thoughts:
I really wish python had ruby `:symbols`, they simplify dsl writing in ruby soooo much. In python you end up having to implement a parser and some degree of an interpreter to accomplish the same... and even then you still have to invoke it with a string. Though in all honesty there is a lot of python metaprogramming facilities that I don't understand so it's likely the answer hides in there. Nevertheless this is much simpler (easier?) in ruby.

pickle comes in handy in a pinch, but it does feel like playing with something I shouldn't. It's ok if you control everything you are pickling and unpickling, but maybe consider alternatives lest I need to change this much later when it'll be harder?